Split files and add docs
