+++ src/idom/getUnit.ts
@@ -1,18 +1,17 @@
 /**
+ * Get unit from string value
  *
- * @param val
+ * ex:
+ *  "200px"
+ *  return "px"
+ *
  */
-export const getUnit = (val: string | number): string | undefined => {
-  if (typeof val === "number") return undefined
-  else {
-    const split =
-      /[+-]?\d*\.?\d+(?:\.\d+)?(?:[eE][+-]?\d+)?(%|px|pt|em|rem|in|cm|mm|ex|ch|pc|vw|vh|vmin|vmax|deg|rad|turn)?$/.exec(
-        val
-      )
-    if (split) return split[1]
-    else return undefined
-  }
-}
+export const getUnit = (value: string | number): string | undefined =>
+  typeof value === "number"
+    ? undefined
+    : /[+-]?\d*\.?\d+(?:\.\d+)?(?:[eE][+-]?\d+)?(%|px|pt|em|rem|in|cm|mm|ex|ch|pc|vw|vh|vmin|vmax|deg|rad|turn)?$/.exec(
+        value
+      )?.[1]
 
 /**
  * Available units
