+++ src/idom/idom.ts
@@ -1,9 +1,10 @@
 import { Interpol } from "../index"
 import { IInterpolConstruct, IUpdateParams } from "../interpol/Interpol"
 import debug from "@wbe/debug"
-import { getUnit } from "./getUnit"
-
+import { extractValueAndUnit } from "./extractValueAndUnit"
 const log = debug(`interpol:idom`)
+
+// ----------------------------------------------------------------------------- TYPES
 
 type AdditionalProperties = {
   x: number | string
@@ -28,94 +29,7 @@ interface ITPOptions<KEYS = any>
 type Styles = Record<keyof CSSStyleDeclaration, number | string> & AdditionalProperties
 type Options = ITPOptions & Styles
 
-/**
- *
- * UTILS
- * - forceUnit
- * - or unit from brut value
- * - or unit from computed value
- *
- * - value from brute value
- * - or value from computed
- */
-export const geValueAndUnit = (
-  $target: HTMLElement,
-  key: string,
-  brutValue: number | string,
-  pUnit?: string,
-  proxyWindow = window
-): [value: number, unit: string] => {
-  const computedValue = proxyWindow.getComputedStyle($target).getPropertyValue(key)
-  const currUnit = getUnit(brutValue) || getUnit(computedValue)
-  const unit = pUnit || currUnit
-  let value =
-    (typeof brutValue === "string" ? parseFloat(brutValue) : brutValue) ?? parseFloat(computedValue)
-  if (pUnit) value = convertValueToUnitValue($target, value, currUnit, pUnit)
-  return [value, unit]
-}
-
-/**
- * Convert a value from unit value to another unit value
- * ex:
- *  [2rem, 50%]
- *  -> [0.5%, 50%]
- *
- * return a value
- */
-function convertValueToUnitValue(
-  el: HTMLElement,
-  value,
-  fromUnit: string,
-  toUnit: string,
-  pWindow = window,
-  pDocument = document
-): number {
-  if (fromUnit === toUnit) {
-    log("unit are the same, return value", value)
-    return value
-  }
-
-  // log({ fromUnit, toUnit, value })
-  // create a temp node element in order to get his width
-  const tempEl = pDocument.createElement(el.tagName)
-  const parentEl = el.parentNode && el.parentNode !== pDocument ? el.parentNode : pDocument.body
-  tempEl.style.position = "absolute"
-  tempEl.style.width = 100 + "%"
-  parentEl.appendChild(tempEl)
-  const tempWidthPx = tempEl.offsetWidth
-  parentEl.removeChild(tempEl)
-
-  // get font-sizes
-  const parentFontSize = parseFloat(pWindow.getComputedStyle(el.parentElement).fontSize)
-  const baseFontSize = parseFloat(pWindow.getComputedStyle(pDocument.documentElement).fontSize)
-
-  // need some tests...
-  const units = {
-    rem: baseFontSize,
-    em: parentFontSize,
-    "%": tempWidthPx / 100,
-    px: 1,
-    pt: 4 / 3,
-    in: 96,
-    cm: 96 / 2.54,
-    mm: 96 / 25.4,
-    ex: parentFontSize / 2,
-    ch: parentFontSize * 8.8984375,
-    pc: 16,
-    vw: pWindow.innerWidth / 100,
-    vh: pWindow.innerHeight / 100,
-    vmin: Math.min(pWindow.innerWidth, pWindow.innerHeight) / 100,
-    vmax: Math.max(pWindow.innerWidth, pWindow.innerHeight) / 100,
-    deg: Math.PI / 180,
-    rad: 1,
-    turn: 360,
-  }
-
-  const pixelValue = value * units[fromUnit]
-  return pixelValue / units[toUnit]
-}
-
-// -----------------------------------------------------------------------------
+// ----------------------------------------------------------------------------- IDOM
 
 /**
  * Interdom
@@ -167,16 +81,15 @@ export function idom(
     if (Array.isArray(value)) {
       const [vFrom, vTo] = value
       hasExplicitFrom = vFrom !== null && vFrom !== undefined
-      to = geValueAndUnit(target, key, vTo)
-      from = geValueAndUnit(target, key, vFrom, to[1])
+      to = extractValueAndUnit(target, key, vTo)
+      from = extractValueAndUnit(target, key, vFrom, to[1])
     }
     // value is a number or a string, not an array
     else {
       const cptValue = window.getComputedStyle(target).getPropertyValue(key)
-      to = geValueAndUnit(target, key, value)
-      from = geValueAndUnit(target, key, cptValue, to[1])
+      to = extractValueAndUnit(target, key, value)
+      from = extractValueAndUnit(target, key, cptValue, to[1])
     }
-
     log(key, { from, to })
 
     // return interpol instance for current key
